Adds TripController
